concurrency: initial draft
Spell checking is expensive.
- If someone adds an additional commit, it will probably invalidate the previous report.
As such, we choose to cancel the earlier run and save CPU cycles for the new run.

Update branch.
- It is unlikely that two people will ask the bot to apply different changes at the same
time, however, it could happen. In general, the bot should do something reasonable as
long as both requests are reasonable (most likely an older one and a newer one).
As such, we choose to queue the later run instead of expecting one of the two to
randomly lose the race and fail.