Reduce scope of excluded content
Previously the entire `.github` directory was excluded, but...
There are lots of files in that tree which are user facing and that tend to
have typos.

Instead, this change adds a couple of sample words to the dictionary and
scopes the exclude to only cover the spelling metadata directory itself.